Move utils function in separate file
