+++ static/src/containers/Sounds/reducer.js
@@ -3,20 +3,11 @@ import { FETCH_SOUNDS_SUCCESS, UPDATE_SOUNDS_POSITION, SELECT_SOUND_BY_ID,
   STOP_AUDIO_SRC, MAP_COMPUTATION_COMPLETE, REMOVE_SOUND }
   from './actions';
 import { UPDATE_MAP_POSITION } from '../Map/actions';
-import { MAP_SCALE_FACTOR } from '../../constants';
+import { computeSoundGlobalPosition } from './utils';
 import sessions from '../Sessions/reducer';
 import soundInfo from '../SoundInfo/reducer';
 
-export const computeSoundGlobalPosition = (tsnePosition, spacePosition, mapPosition) => {
-  const { translateX, translateY, scale } = mapPosition;
-  const cx = ((tsnePosition.x + (window.innerWidth / (MAP_SCALE_FACTOR * 2))) *
-    MAP_SCALE_FACTOR * scale * spacePosition.x) + translateX;
-  const cy = ((tsnePosition.y + (window.innerHeight / (MAP_SCALE_FACTOR * 2))) *
-    MAP_SCALE_FACTOR * scale * spacePosition.y) + translateY;
-  return { cx, cy };
-};
-
-const sound = (state, action) => {
+export const sound = (state, action) => {
   const soundID = state.id;
   switch (action.type) {
     case UPDATE_SOUNDS_POSITION: {
@@ -48,7 +39,7 @@ const sound = (state, action) => {
   }
 };
 
-const byID = (state = {}, action) => {
+export const byID = (state = {}, action) => {
   switch (action.type) {
     case FETCH_SOUNDS_SUCCESS: {
       return Object.assign({}, state, action.sounds);
@@ -94,7 +85,7 @@ const byID = (state = {}, action) => {
   }
 };
 
-const selectedSound = (state = 0, action) => {
+export const selectedSound = (state = 0, action) => {
   switch (action.type) {
     case SELECT_SOUND_BY_ID:
       return action.soundID || 0;
