+++ static/src/containers/Spaces/reducer.js
@@ -2,7 +2,7 @@ import { REMOVE_SPACE } from './actions';
 import { UPDATE_MAP_POSITION } from '../Map/actions';
 import { FETCH_SOUNDS_REQUEST, FETCH_SOUNDS_SUCCESS, FETCH_SOUNDS_FAILURE }
   from '../Sounds/actions';
-import { computeSoundGlobalPosition } from '../Sounds/reducer';
+import { computeSoundGlobalPosition } from '../Sounds/utils';
 import { getMapCenter } from '../../utils/uiUtils';
 import { generateListOfSpacesOriginPositions } from '../../utils/misc';
 import sessions from '../Sessions/reducer';
